Move dnstwist into it's own folder and fix pe pytest
