+++ backend/pe/src/pe_source/dnstwist/__init__.py
@@ -1,0 +1,5 @@
+"""DNSTwist scan source."""
+
+from pe_source.dnstwist.dnstwist import checkBlocklist, run_dnstwist
+
+__all__ = ["checkBlocklist", "run_dnstwist"]
